refactor mock tests
